WIP: Retype int->uint in various implementations

+++ go/cell.go
@@ -1,14 +1,14 @@
 package main
 
 type Cell struct {
-  x int
-  y int
+  x uint32
+  y uint32
   alive bool
   next_state bool
   neighbours []*Cell
 }
 
-func new_cell(x int, y int, alive bool) *Cell {
+func new_cell(x uint32, y uint32, alive bool) *Cell {
   var cell = new(Cell)
   cell.x = x
   cell.y = y
@@ -25,11 +25,9 @@ func (cell *Cell) to_char() rune {
   }
 }
 
-// Implement first using filter/lambda if available. Then implement
-// foreach and for. Use whatever implementation runs the fastest
-func (cell *Cell) alive_neighbours() int {
-  // The following was the fastest method
-  alive_neighbours := 0
+func (cell *Cell) alive_neighbours() uint32 {
+  // The following is the fastest
+  alive_neighbours := uint32(0)
   for _, neighbour := range cell.neighbours {
     if neighbour.alive {
       alive_neighbours++
@@ -37,9 +35,10 @@ func (cell *Cell) alive_neighbours() int {
   }
   return alive_neighbours
 
-  // The following also works but is slower
-  // alive_neighbours := 0
-  // for i := range(len(cell.neighbours)) {
+  // The following is slower
+  // alive_neighbours := uint32(0)
+  // count := len(cell.neighbours)
+  // for i := range(count) {
   //   neighbour := cell.neighbours[i]
   //   if neighbour.alive {
   //     alive_neighbours++
